Todo wrt the (hacked) reboot command
The current implementation of the Apex 'redboot' command is not ideal
and there are several improvements that can be done. Note what those are
in case more interest is gathered around this topic in the future.

OTOH, the current implementation is enough to make TFTP+NFS NetBSD
booting possible by just powering the device (after making the kernel
available through TFTP and a root file system over NFS).

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>